fix jest tests
